assa

+++ resources/views/livewire/doctor-batch-payment/doctor-batch-paid-list.blade.php
@@ -5,15 +5,21 @@
     <table class="table table-sm table-bordered table-hover">
         <thead class="text-xs bg-info">
             <tr>
-                <th class="col-1">Pay Bill Date</th>
-                <th class="col-1 text-right">Pay Bill Amt.</th>
                 <th class="col-1">Pay Bill No.</th>
+
+                <th class="col-1">Pay Bill Date</th>
+                <th class="col-1 text-right">Pay Paid.</th>
+                <th class="col-1 text-right">WTax.</th>
+                <th class="col-1 text-right">Total.</th>
                 <th class="col-1 text-center">No. of Patient</th>
+                <th class="col-1 text-center bg-warning">Status</th>
                 <th class="col-1 bg-secondary">OR No.</th>
                 <th class="col-1 bg-secondary">OR Date</th>
                 <th class="col-1 bg-secondary">From Date</th>
                 <th class="col-1 bg-secondary">To Date</th>
-                <th class="col-1 bg-info text-center">Action</th>
+                @can('patient.doctor.batch.delete')
+                    <th class="col-1 bg-info text-center">Action</th>
+                @endcan
             </tr>
         </thead>
         <tbody class="text-xs">
@@ -23,19 +29,25 @@
                         <a target="_BLANK"
                             href="{{ route('vendorsbill_payment_edit', ['id' => $list->ID]) }}">{{ $list->CODE }}</a>
                     </td>
-                    <td class="text-right">{{ number_format($list->AMOUNT, 2) }}</td>
+
                     <td>{{ \Carbon\Carbon::parse($list->DATE)->format('m/d/Y') }} </td>
+                    <td class="text-right">{{ number_format($list->AMOUNT, 2) }}</td>
+                    <td class="text-right">{{ number_format($list->TAX_AMOUNT, 2) }}</td>
+                    <td class="text-right">{{ number_format($list->AMOUNT + $list->TAX_AMOUNT, 2) }}</td>
                     <td class="text-center"> {{ $list->TOTAL_COUNT }} </td>
+                    <td class="text-center"> {{ $list->STATUS }} </td>
                     <td> {{ $list->OR_NO }}</td>
                     <td>{{ \Carbon\Carbon::parse($list->OR_DATE)->format('m/d/Y') }} </td>
                     <td>{{ \Carbon\Carbon::parse($list->FROM_DATE)->format('m/d/Y') }} </td>
                     <td>{{ \Carbon\Carbon::parse($list->TO_DATE)->format('m/d/Y') }} </td>
-                    <td>
-                        <button type="button" class="btn btn-danger btn-xs text-xs w-100"
-                            wire:click='deleteItem({{ $list->ID }})' wire:confirm='Are you sure to deleted?'>
-                            <i class="fa fa-trash" aria-hidden="true"></i>
-                        </button>
-                    </td>
+                    @can('patient.doctor.batch.delete')
+                        <td>
+                            <button type="button" class="btn btn-danger btn-xs text-xs w-100"
+                                wire:click='deleteItem({{ $list->ID }})' wire:confirm='Are you sure to deleted?'>
+                                <i class="fa fa-trash" aria-hidden="true"></i>
+                            </button>
+                        </td>
+                    @endcan
                 </tr>
             @endforeach
         </tbody>
